Solved SIGPIPE bug
I think I have solved the SIGPIPE bug. Previously, read_from_sock would
have a timeout in select() - this was not very forgiving timeout. If the
client could not push the next chunk of data within 0.01 seconds,
read_from_sock() would finish and close() the sock. I upped the timeout
to 1 second, since we can afford it. If we need to lower the timeout, we
can. This is worth messing with at a later date.